gtools-0.8.3 (2017-10-28); faster `gisid` when data is sorted
Features

- `gisid` includes an internal check to see if the data is sorted.
  This means that if there are two duplicate rows in unsorted data
  or if the data is already sorted, `gisid` will give a result much
  faster. However, if the data is not sorted it will be marginally
  slower as it will execute the rest of the code normally.

Backwards Incompatible

- `gisid` and `hashsort` are no longer rclass. Both will exit early if
  the data is already sorted, and `gisid` will also exit if it finds a
  duplicate row during the sorted check. Hence they will not always
  store results, making them inconsistent. It would be bad practice
  to continue to have them store `rclass` results.

Enhancements

- Cleaned up the C base somewhat. Improved modularity.
- All `int` were changed to `size_t` or `int64_t` as applicable, since
  `int` is not necessarily aliased to a 64-bit integer on all platforms.

+++ lib/spt-0.2/st_gentools.c
@@ -85,10 +85,10 @@ size_t sf_scalar_int (char *st_scalar, size_t *sval)
  */
 int sf_get_vector_length (char *st_matrix)
 {
-    int ncol = SF_col(st_matrix);
-    int nrow = SF_row(st_matrix);
+    size_t ncol = SF_col(st_matrix);
+    size_t nrow = SF_row(st_matrix);
     if ( (ncol > 1) & (nrow > 1) ) {
-        sf_errprintf("tried to get the length a %d by %d matrix\n", nrow, ncol);
+        sf_errprintf("tried to get the length a %lu by %lu matrix\n", nrow, ncol);
         return (-1);
     }
     return ( ncol > nrow? ncol: nrow );
@@ -105,11 +105,11 @@ int sf_get_vector (char *st_matrix, double *v)
 {
     ST_retcode rc = 0;
 
-    int i;
-    int ncol = SF_col(st_matrix);
-    int nrow = SF_row(st_matrix);
+    size_t i;
+    size_t ncol = SF_col(st_matrix);
+    size_t nrow = SF_row(st_matrix);
     if ( (ncol > 1) & (nrow > 1) ) {
-        sf_errprintf("tried to read a %d by %d matrix into an array\n", nrow, ncol);
+        sf_errprintf("tried to read a %lu by %lu matrix into an array\n", nrow, ncol);
         return(198);
     }
 
@@ -141,15 +141,15 @@ int sf_get_vector_int (char *st_matrix, size_t *v)
     ST_double z;
     ST_retcode rc = 0;
 
-    int i;
-    int ncol = SF_col(st_matrix);
-    int nrow = SF_row(st_matrix);
+    size_t i;
+    size_t ncol = SF_col(st_matrix);
+    size_t nrow = SF_row(st_matrix);
     if ( (ncol > 1) & (nrow > 1) ) {
-        sf_errprintf("tried to read a %d by %d matrix into an array\n", nrow, ncol);
+        sf_errprintf("tried to read a %lu by %lu matrix into an array\n", nrow, ncol);
         return (198);
     }
     else if ( (ncol == 0) & (nrow == 0) ) {
-        sf_errprintf("tried to read a %d by %d matrix into an array\n", nrow, ncol);
+        sf_errprintf("tried to read a %lu by %lu matrix into an array\n", nrow, ncol);
         return (198);
     }
